created new schematree.jsx file, testing for implementation
