+++ client/Components/TestQuery.jsx
@@ -8,39 +8,30 @@ class TestQuery extends Component {
   };
 
   render() {
-    // const schemaTreeData = [
-    //   {
-    //     name: 'Query',
-    //     attributes: {
-    //       type: 'Root'
-    //     },
-    //   },
-    //   children:[ {}]
-    // ];
-    console.log('schema within testquery of Query:', this.props.schema);
-    console.log('RootQuery within testquery of Query:', this.props.schema.Query);
+    // console.log('schema within testquery of Query:', this.props.schema);
+    // console.log('RootQuery within testquery of Query:', this.props.schema.Query);
 
-    function convertJSON(object) {
-      for (let key in object) {
-        if (typeof(object[key]) == "object" && object[key].type !== "InputObjectTypeDefinition") {
-          let obj = convertJSON(object[key]);
-          obj.key = obj.name ? obj.name : "";
-          if (obj.key === "") {
-            obj.name = key;
-          } else {
-            obj.name = obj.key;
-          }
-          if (!object["children"])
-            object["children"] = [];
-          object["children"].push(obj);
-          delete object[key];
-        }
-      }
-      return object;
-    }
+    // function convertJSON(object) {
+    //   for (let key in object) {
+    //     if (typeof(object[key]) == "object" && object[key].type !== "InputObjectTypeDefinition") {
+    //       let obj = convertJSON(object[key]);
+    //       obj.key = obj.name ? obj.name : "";
+    //       if (obj.key === "") {
+    //         obj.name = key;
+    //       } else {
+    //         obj.name = obj.key;
+    //       }
+    //       if (!object["children"])
+    //         object["children"] = [];
+    //       object["children"].push(obj);
+    //       delete object[key];
+    //     }
+    //   }
+    //   return object;
+    // }
     
-    let queryTreeData = convertJSON(this.props.schema);
-    console.log(queryTreeData);
+    // let queryTreeData = convertJSON(this.props.schema);
+    // console.log(queryTreeData);
 
     return (
       <div className="writeQueryBox">
@@ -64,9 +55,9 @@ class TestQuery extends Component {
           
           <div className= "test-query-select">
 
-          <div id="treeWrapper" style={{width: '50em', height: '20em'}}>
+          {/* <div id="treeWrapper" style={{width: '50em', height: '20em'}}>
               <Tree data={queryTreeData} />
-            </div>
+            </div> */}
           {/* <Form.Group as={Col} controlId="select-query"> */}
             {/* <Form.Label column sm={6}>
               Select Query:
