+++ client/public/styling/index.css
@@ -371,19 +371,25 @@ h1 {
     }
 
 .testInput {
-  grid-column: 3;
+  grid-column: 2;
   grid-row: 2;
   height: 100%;
   width: 100%;
 }
 
 .testQuery {
-  grid-column: 2/4;
+  grid-column: 2;
   grid-row: 3;
 }
 
+.queryVisualizer {
+  grid-column: 3;
+  grid-row: 2/5;
+  width: 100%;
+  height: 100%;
+}
 .generateTest {
-  grid-column: 2/4;
+  grid-column: 2;
   grid-row: 4;
   /* background-color: whitesmoke; */
 }
@@ -472,16 +478,16 @@ h1 {
 }
 
   .test-query-write {
-    grid-column: 1;
+    grid-column: 1/3;
     width: 100%;
-    display: inline-block;
+    /* display: inline-block; */
   }
 
-  .test-query-select {
+  /* .test-query-select {
     grid-column: 2;
     width: 100%;
     display: inline-block;
-  }
+  } */
 
 .generateTestBox {
   font-family: 'Poppins', sans-serif;
